+++ packages/contracts/contracts/libs/TreeUtils.sol
@@ -5,12 +5,11 @@ import "./Types.sol";
 
 // helpers for converting to/from field elems, uint256s, and/or bytes, and hashing them
 library TreeUtils {
-    uint256 public constant DEPTH = 32;
+    uint256 public constant DEPTH = 16;
     uint256 public constant BATCH_SIZE = 16;
-    uint256 public constant LOG2_BATCH_SIZE = 4;
-    uint256 public constant BATCH_SUBTREE_DEPTH = 4;
+    uint256 public constant BATCH_SUBTREE_DEPTH = 2;
     uint256 public constant EMPTY_TREE_ROOT =
-        21443572485391568159800782191812935835534334817699172242223315142338162256601;
+        9533201250583817767896570092866591469094150406835227552485691564931228351592;
 
     // packs a field element for the `encodedPathAndHash` input to the subtree update verifier
     // `subtreeIdx` is the index of the subtree's leftmost element in the tree
@@ -23,10 +22,13 @@ library TreeUtils {
             subtreeIdx % BATCH_SIZE == 0,
             "subtreeIdx not multiple of BATCH_SIZE"
         );
-        uint256 encodedPathAndHash = uint256(subtreeIdx) >> BATCH_SUBTREE_DEPTH;
+
+        // we shift by 2 * depth because the tree is quaternary
+        uint256 encodedPathAndHash = uint256(subtreeIdx) >>
+            (2 * BATCH_SUBTREE_DEPTH);
         encodedPathAndHash |=
             accumulatorHashHi <<
-            (DEPTH - BATCH_SUBTREE_DEPTH);
+            (2 * (DEPTH - BATCH_SUBTREE_DEPTH));
 
         return encodedPathAndHash;
     }