+++ packages/contracts/contracts/test/unit/OffchainMerkleTree.t.sol
@@ -7,9 +7,9 @@ import {ParseUtils} from "../utils/ParseUtils.sol";
 import {TreeUtils} from "../../libs/TreeUtils.sol";
 import {TreeTest, TreeTestLib} from "../utils/TreeTest.sol";
 import {LibOffchainMerkleTree, OffchainMerkleTree} from "../../libs/OffchainMerkleTree.sol";
-import {IHasherT3, IHasherT6} from "../interfaces/IHasher.sol";
+import {IHasherT3, IHasherT5, IHasherT6} from "../interfaces/IHasher.sol";
 import {ISubtreeUpdateVerifier} from "../../interfaces/ISubtreeUpdateVerifier.sol";
-import {PoseidonHasherT3, PoseidonHasherT6} from "../utils/PoseidonHashers.sol";
+import {PoseidonHasherT3, PoseidonHasherT5, PoseidonHasherT6} from "../utils/PoseidonHashers.sol";
 import {PoseidonDeployer} from "../utils/PoseidonDeployer.sol";
 import {TestSubtreeUpdateVerifier} from "../harnesses/TestSubtreeUpdateVerifier.sol";
 import "../../libs/Types.sol";
@@ -21,12 +21,16 @@ contract TestOffchainMerkleTree is PoseidonDeployer {
     OffchainMerkleTree merkle;
     ISubtreeUpdateVerifier subtreeUpdateVerifier;
     IHasherT3 hasherT3;
+    IHasherT5 hasherT5;
     IHasherT6 hasherT6;
     TreeTest treeTest;
 
     event InsertNoteCommitments(uint256[] commitments);
 
     event InsertNote(EncodedNote note);
+
+    uint256 constant DEPTH_TO_SUBTREE =
+        TreeUtils.DEPTH - TreeUtils.BATCH_SUBTREE_DEPTH;
 
     function setUp() public virtual {
         // Deploy poseidon hasher libraries
@@ -35,8 +39,9 @@ contract TestOffchainMerkleTree is PoseidonDeployer {
             new TestSubtreeUpdateVerifier()
         );
         hasherT3 = IHasherT3(new PoseidonHasherT3(poseidonT3));
+        hasherT5 = IHasherT5(new PoseidonHasherT5(poseidonT5));
         hasherT6 = IHasherT6(new PoseidonHasherT6(poseidonT6));
-        treeTest.initialize(hasherT3, hasherT6);
+        treeTest.initialize(hasherT3, hasherT5, hasherT6);
         merkle.initialize(address(subtreeUpdateVerifier));
     }
 
@@ -49,31 +54,31 @@ contract TestOffchainMerkleTree is PoseidonDeployer {
         );
 
         // test that hashing empty batch total gives EMPTY_TREE_ROOT
-        uint256[] memory path = treeTest.computeInitialRoot(batch);
-        assertEq(path[path.length - 1], TreeUtils.EMPTY_TREE_ROOT);
+        uint256[][3] memory path = treeTest.computeInitialPaths(batch);
+        assertEq(path[0][DEPTH_TO_SUBTREE], TreeUtils.EMPTY_TREE_ROOT);
 
-        // test computeInitialRoot for non-empty batch
+        // test computeInitialPaths for non-empty batch
         batch = new uint256[](2);
         batch[0] = 420;
         batch[1] = 69;
-        path = treeTest.computeInitialRoot(batch);
+        path = treeTest.computeInitialPaths(batch);
         assertEq(
-            path[path.length - 1],
-            7535458605132084619456785809582285707117893595742786994560375527566624811343
+            path[0][DEPTH_TO_SUBTREE],
+            1876908926272187450761598598188752060951923126538985491623069278962090072418
         );
-        assertEq(path[0], treeTest.computeSubtreeRoot(batch));
+        assertEq(path[0][0], treeTest.computeSubtreeRoot(batch));
 
-        // test computeNewRoot for non-empty batch
+        // test computeNewPaths for non-empty batch
         batch = new uint256[](3);
         batch[0] = 9;
         batch[1] = 1;
         batch[2] = 1449;
-        path = treeTest.computeNewRoot(batch, path, 16);
+        path = treeTest.computeNewPaths(batch, path, 16);
         assertEq(
-            path[path.length - 1],
-            6984220783167935932287489881196784031196766582157979071264100186030762206286
+            path[0][DEPTH_TO_SUBTREE],
+            7121119364410700771071060140874401316335848409647514241451685854290541005256
         );
-        assertEq(path[0], treeTest.computeSubtreeRoot(batch));
+        assertEq(path[0][0], treeTest.computeSubtreeRoot(batch));
     }
 
     function testInsertSingleNote() public {
@@ -98,8 +103,8 @@ contract TestOffchainMerkleTree is PoseidonDeployer {
         assertEq(merkle.getRoot(), TreeUtils.EMPTY_TREE_ROOT);
 
         // compute new root and call `applySubtreeUpdate`
-        uint256[] memory path = treeTest.computeInitialRoot(batch);
-        uint256 newRoot = path[path.length - 1];
+        uint256[][3] memory path = treeTest.computeInitialPaths(batch);
+        uint256 newRoot = path[0][DEPTH_TO_SUBTREE];
         merkle.applySubtreeUpdate(newRoot, dummyProof());
 
         assertEq(uint256(merkle.getCount()), 16);
@@ -122,8 +127,8 @@ contract TestOffchainMerkleTree is PoseidonDeployer {
         assertEq(merkle.getRoot(), TreeUtils.EMPTY_TREE_ROOT);
 
         // apply subtree update
-        uint256[] memory path = treeTest.computeInitialRoot(batch);
-        uint256 newRoot = path[path.length - 1];
+        uint256[][3] memory path = treeTest.computeInitialPaths(batch);
+        uint256 newRoot = path[0][DEPTH_TO_SUBTREE];
         merkle.applySubtreeUpdate(newRoot, dummyProof());
 
         assertEq(merkle.getCount(), 16);
@@ -143,7 +148,7 @@ contract TestOffchainMerkleTree is PoseidonDeployer {
         }
 
         // Insert 9 ncs
-        uint256[] memory ncs = new uint256[](10);
+        uint256[] memory ncs = new uint256[](9);
         for (uint256 i = 0; i < 9; i++) {
             ncs[i] = nc;
         }
@@ -158,26 +163,28 @@ contract TestOffchainMerkleTree is PoseidonDeployer {
             batch[i] = nc;
         }
 
-        uint256[] memory path = treeTest.computeInitialRoot(batch);
-        uint256 _newRoot = path[path.length - 1];
+        uint256[][3] memory path = treeTest.computeInitialPaths(batch);
+        uint256 _newRoot = path[0][DEPTH_TO_SUBTREE];
 
-        uint256 newRoot = 2148530186383747530821653986434349341874407543492575165183948509644419849075;
+        uint256 newRoot = 7692074189416150984263394138680301006228090185414219994496061801546117138813;
 
         assertEq(newRoot, _newRoot);
 
+        uint256 accumulatorHash = merkle.getAccumulatorHash();
+        console.log("accumulatorHash", accumulatorHash);
+        (uint256 hi, uint256 lo) = TreeUtils.uint256ToFieldElemLimbs(
+            accumulatorHash
+        );
+
+        console.log("accumulatorHashHi", hi);
+
         uint256[] memory pis = merkle._calculatePublicInputs(newRoot);
-        assertEq(
-            pis[0],
-            21443572485391568159800782191812935835534334817699172242223315142338162256601
-        );
-        assertEq(
-            pis[1],
-            2148530186383747530821653986434349341874407543492575165183948509644419849075
-        );
-        assertEq(pis[2], 1073741824);
+        assertEq(pis[0], TreeUtils.EMPTY_TREE_ROOT);
+        assertEq(pis[1], newRoot);
+        assertEq(pis[2], 1879048192);
         assertEq(
             pis[3],
-            4369603618049527331146763788699655885066799772437199894124267345370035181946
+            13761535849878919798310125019909519451162264697046676736248712268787268459921
         );
     }
 
@@ -188,8 +195,8 @@ contract TestOffchainMerkleTree is PoseidonDeployer {
 
     function dummyNote() internal pure returns (EncodedNote memory) {
         EncodedNote memory note = EncodedNote({
-            ownerH1: 16114171923265390730037465875328827721281782660087141077700479736598096658937,
-            ownerH2: 10977258428915190383432832691667013955459124698254120657094471191004412212417,
+            ownerH1: 20053872845712750666020333248434368879858874000328815279916175647306793909806,
+            ownerH2: 10878178814994881930842668029692572520203302021151403528591159382456948662398,
             nonce: 1,
             encodedAssetAddr: 917551056842671309452305380979543736893630245704,
             encodedAssetId: 5,